Split oversized test files; centralize mockMode setup
