+++ apps/control-plane/src/test/channel-crud.test.ts
@@ -8,8 +8,6 @@ import { resetDb } from "./helpers/reset-db.js";
 import { createAuthCookie } from "./helpers/auth.js";
 import { bootstrap as bootstrapHub } from "./helpers/bootstrap.js";
 
-config.discordBridge.mockMode = true;
-
 beforeEach(async () => {
   if (pool) {
     await initDb();
